refactor(components): consolidate 3 article cards into ArticleCard with size prop
Before: HeroArticle (161) + SecondaryArticle (136) + TertiaryArticle (134)
totaling 431 lines, with 39-line clones between all pairs per jscpd.
After: ArticleCard (63) + ArticleCardBody (270) totaling 333 lines, with
a size: 'hero' | 'secondary' | 'tertiary' prop driving size-indexed class
lookups and three targeted structural branches.

Net deletion: 98 lines.
Updated 3 call sites in CoverSection.svelte (one file).
Visual output preserved — same DOM, classes, styles, behavior.

Body extracted into ArticleCardBody.svelte because Svelte 4's template
parser forbids opening an element inside one {#if} block and closing it
in a later block — required for the hero-only inner flex wrapper. A
single-file consolidation would have required duplicating ~150 lines of
body markup or dropping hero's inner wrapper (changing DOM depth). The
split keeps ArticleCard as the public API while avoiding both.
